refactor: customKeyGenerator을 클래스 단위에서 적용하도록 모든 캐시에 적용
