+++ backend/src/main/java/reviewme/question/repository/OptionItemRepository.java
@@ -1,6 +1,7 @@
 package reviewme.question.repository;
 
 import java.util.List;
+import org.springframework.cache.annotation.CacheConfig;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -9,19 +10,20 @@ import reviewme.question.domain.OptionItem;
 import reviewme.question.domain.OptionType;
 
 @Repository
+@CacheConfig(cacheNames = "templateCache", keyGenerator = "cacheKeyGenerator")
 public interface OptionItemRepository extends JpaRepository<OptionItem, Long> {
 
-    @Cacheable(value = "templateCache", keyGenerator = "customKeyGenerator")
+    @Cacheable
     List<OptionItem> findAllByOptionGroupId(long optionGroupId);
 
-    @Cacheable(value = "templateCache", keyGenerator = "customKeyGenerator")
+    @Cacheable
     @Query("""
             SELECT o FROM OptionItem o
             WHERE o.optionType = :optionType
             """)
     List<OptionItem> findAllByOptionType(OptionType optionType);
 
-    @Cacheable(value = "templateCache", keyGenerator = "customKeyGenerator")
+    @Cacheable
     @Query("""
             SELECT o FROM OptionItem o
             JOIN OptionGroup og ON o.optionGroupId = og.id
@@ -30,7 +32,7 @@ public interface OptionItemRepository extends JpaRepository<OptionItem, Long> {
             """)
     List<OptionItem> findAllOptionItemsByIdOrderByPosition(long questionId);
 
-    @Cacheable(value = "templateCache", keyGenerator = "customKeyGenerator")
+    @Cacheable
     @Query("""
             SELECT o FROM OptionItem o
             JOIN OptionGroup og ON o.optionGroupId = og.id
