+++ backend/src/main/java/reviewme/question/repository/QuestionRepository.java
@@ -4,6 +4,7 @@ import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
+import org.springframework.cache.annotation.CacheConfig;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -11,9 +12,10 @@ import org.springframework.stereotype.Repository;
 import reviewme.question.domain.Question;
 
 @Repository
+@CacheConfig(cacheNames = "templateCache", keyGenerator = "cacheKeyGenerator")
 public interface QuestionRepository extends JpaRepository<Question, Long> {
 
-    @Cacheable(value = "templateCache", key = "#templateId")
+    @Cacheable
     @Query("""
             SELECT q.id FROM Question q
             JOIN SectionQuestion sq
@@ -24,7 +26,7 @@ public interface QuestionRepository extends JpaRepository<Question, Long> {
             """)
     Set<Long> findAllQuestionIdByTemplateId(long templateId);
 
-    @Cacheable(value = "templateCache", key = "#templateId")
+    @Cacheable
     @Query("""
             SELECT q FROM Question q
             JOIN SectionQuestion sq
@@ -35,7 +37,7 @@ public interface QuestionRepository extends JpaRepository<Question, Long> {
             """)
     List<Question> findAllByTemplatedId(long templateId);
 
-    @Cacheable(value = "templateCache", key = "#sectionId")
+    @Cacheable
     @Query("""
             SELECT q FROM Question q
             JOIN SectionQuestion sq ON q.id = sq.questionId
@@ -44,13 +46,13 @@ public interface QuestionRepository extends JpaRepository<Question, Long> {
             """)
     List<Question> findAllBySectionIdOrderByPosition(long sectionId);
 
-    @Cacheable(value = "templateCache", key = "#questionIds.hashCode()")
+    @Cacheable
     @Query("""
         SELECT q FROM Question q
         WHERE q.id IN :questionIds
         """)
     Collection<Question> findAllById(Collection<Long> questionIds);
 
-    @Cacheable(value = "templateCache", key = "#questionId")
+    @Cacheable
     Optional<Question> findById(long questionId);
 }
