+++ backend/src/main/java/reviewme/template/repository/SectionRepository.java
@@ -2,6 +2,7 @@ package reviewme.template.repository;
 
 import java.util.List;
 import java.util.Optional;
+import org.springframework.cache.annotation.CacheConfig;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -9,9 +10,10 @@ import org.springframework.stereotype.Repository;
 import reviewme.template.domain.Section;
 
 @Repository
+@CacheConfig(cacheNames = "templateCache", keyGenerator = "cacheKeyGenerator")
 public interface SectionRepository extends JpaRepository<Section, Long> {
 
-    @Cacheable(value = "templateCache", key = "#templateId")
+    @Cacheable
     @Query("""
             SELECT s FROM Section s
             JOIN TemplateSection ts
@@ -21,7 +23,7 @@ public interface SectionRepository extends JpaRepository<Section, Long> {
             """)
     List<Section> findAllByTemplateId(long templateId);
 
-    @Cacheable(value = "templateCache", key = "#sectionId + '_' + #templateId")
+    @Cacheable
     @Query("""
             SELECT s FROM Section s
             JOIN TemplateSection ts ON s.id = ts.sectionId
